+++ src/components/panel/panel.js
@@ -18,6 +18,7 @@ export const EuiPanel = ({
   hasShadow,
   grow,
   panelRef,
+  onClick,
   ...rest
 }) => {
 
@@ -27,18 +28,22 @@ export const EuiPanel = ({
     {
       'euiPanel--shadow': hasShadow,
       'euiPanel--flexGrowZero': !grow,
+      'euiPanel--isClickable': onClick,
     },
     className
   );
 
+  const PanelTag = onClick ? 'button' : 'div';
+
   return (
-    <div
+    <PanelTag
+      onClick={onClick}
       ref={panelRef}
       className={classes}
       {...rest}
     >
       {children}
-    </div>
+    </PanelTag>
   );
 
 };
@@ -50,6 +55,7 @@ EuiPanel.propTypes = {
   paddingSize: PropTypes.oneOf(SIZES),
   grow: PropTypes.bool,
   panelRef: PropTypes.func,
+  onClick: PropTypes.func,
 };
 
 EuiPanel.defaultProps = {